requirements.txt
fastapi
uvicorn
python-dotenv
authlib
httpx
itsdangerous
